Modernize (typescript, hooks, vite, vitest, @testing-library) (#8)
I found that this package is pretty useful, but I wanted to use it in my typescript + react-hooks based project and preferred not to have a class-based component dependency (and also wanted to avoid prop-types as a dependency, even though maybe it gets tree-shaken out).  It was small enough that I rewrote it for myself, and figured I might as well submit a PR in case you're interested in updating some of the dependencies.