Don't emit js by default (#10)
The lack of a specified output directory causes the executing of `npx tsc` (useful command for typechecking) to generate `.js` and `.d.ts` files in the `src` directory. This is convenient to deal with.

This PR changes the behavior of tsconfig to not emit anything by default, and let's the commands in `package.json` override that behavior.